+++ animations/lib/main.dart
@@ -11,7 +11,7 @@ import 'src/basics/04_tweens.dart';
 import 'src/basics/05_animated_builder.dart';
 import 'src/basics/06_custom_tween.dart';
 import 'src/basics/07_tween_sequence.dart';
-import 'src/basics/08_animated_list.dart';
+import 'src/misc/animated_list.dart';
 import 'src/misc/card_swipe.dart';
 import 'src/misc/carousel.dart';
 import 'src/misc/expand_card.dart';
@@ -58,10 +58,6 @@ final basicDemos = [
       name: 'Tween Sequences',
       route: TweenSequenceDemo.routeName,
       builder: (context) => TweenSequenceDemo()),
-  Demo(
-      name: 'AnimatedList',
-      route: AnimatedListDemo.routeName,
-      builder: (context) => AnimatedListDemo()),
 ];
 
 final miscDemos = [
@@ -89,6 +85,10 @@ final miscDemos = [
       name: 'Spring Physics',
       route: PhysicsCardDragDemo.routeName,
       builder: (context) => PhysicsCardDragDemo()),
+  Demo(
+      name: 'AnimatedList',
+      route: AnimatedListDemo.routeName,
+      builder: (context) => AnimatedListDemo()),
 ];
 
 final basicDemoRoutes =
